refactor: move the member reassign calculation outside of the modal controller
